Move default converter to tests
