+++ pulsar-broker/src/test/java/org/apache/pulsar/client/converter/PayloadConverterTest.java
@@ -19,7 +19,6 @@
 package org.apache.pulsar.client.converter;
 
 import com.google.common.collect.Sets;
-
 import io.netty.buffer.ByteBuf;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -36,7 +35,6 @@ import org.apache.pulsar.client.api.ProducerConsumerBase;
 import org.apache.pulsar.client.api.Schema;
 import org.apache.pulsar.client.api.SubscriptionInitialPosition;
 import org.apache.pulsar.client.api.TopicMetadata;
-import org.apache.pulsar.client.impl.DefaultPayloadConverter;
 import org.apache.pulsar.common.api.proto.MessageMetadata;
 import org.apache.pulsar.common.policies.data.ClusterData;
 import org.apache.pulsar.common.policies.data.TenantInfoImpl;
